sw_services: xilplmi: Fixes for MISRA C violation Rule 10.4
MISRA C Rule 10.4 States: Both operands of an operator in which the usual
arithmetic conversionsare performed shall have the same essential type category

This patch fixes MISRA C Rule 10.4 violation by using the same type variable
for both operands where possible


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_err.c
@@ -515,7 +515,7 @@ void XPlmi_ErrIntrHandler(void *CallbackRef)
 				(ErrorTable[Index].Handler != NULL) &&
 				(ErrorTable[Index].Action != XPLMI_EM_ACTION_NONE)) {
 				/* PSM errors are handled in PsmErrHandler */
-				if (Index != XPLMI_NODEIDX_ERROR_PMC_PSM_NCR) {
+				if (Index != (u32)XPLMI_NODEIDX_ERROR_PMC_PSM_NCR) {
 				      XPlmi_EmDisable(
 					XPLMI_EVENT_ERROR_PMC_ERR1, Index);
 				}
